localize shallow water to more languages
I only know Russian, English and Japanese, so I did this with the help
of DeepL. I did my best to make sure it's correct, but I may have made
some mistakes.

For most languages, I didn't write a description for the new setting as
I have no confidence in getting that right.